Add `slack-display-name` to the JWT claims when a user is logged in by Slack. It's the logged in Slack org's user name for them.
Store `display-name` in the Slack user in the DB for their Slack login(s).

Add `title` to the user schema. We'll eventually use it for People feature. Pickup `title` and `timezone` from Slack when they login via Slack.

Add `slack-display-name` to the team roster and improve how we parse names from Slack.